# docs/swagger.yaml
openapi: 3.0.0
info:
  title: GoHealth API Documentation
  version: 1.0.0
  description: Complete API documentation for GoHealth application. A comprehensive health and fitness tracking API that provides endpoints for managing user profiles, BMI calculations, meal tracking, and activity tracking.
  contact:
    name: GoHealth Team
    email: support@gohealth.example.com
servers:
  - url: http://localhost:3000/api
    description: Local development server
  - url: https://api.gohealth.example.com/api
    description: Production server
  - url: https://staging-api.gohealth.example.com/api
    description: Staging server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        email:
          type: string
          format: email
          example: user@example.com
        name:
          type: string
          example: John Doe
        age:
          type: integer
          example: 30
        gender:
          type: string
          enum: [MALE, FEMALE]
          example: MALE
        height:
          type: number
          format: float
          example: 175.5
        weight:
          type: number
          format: float
          example: 70.2
        activityLevel:
          type: string
          enum: [SEDENTARY, LIGHTLY, ACTIVE, MODERATELY_ACTIVE, VERY_ACTIVE, EXTRA_ACTIVE]
          example: ACTIVE
        profileImage:
          type: string
          nullable: true
          example: https://example.com/profile.jpg
        createdAt:
          type: string
          format: date-time
          example: 2023-01-01T00:00:00.000Z
        updatedAt:
          type: string
          format: date-time
          example: 2023-01-01T00:00:00.000Z
    AuthResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
          example: Authentication successful
        data:
          type: object
          properties:
            user:
              $ref: '#/components/schemas/User'
            accessToken:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            refreshToken:
              type: string
              example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
            tokenType:
              type: string
              example: Bearer
            expiresIn:
              type: string
              example: 1h
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          example: Error message
        error:
          type: object
          properties:
            status:
              type: integer
              example: 400
            name:
              type: string
              example: BadRequestError
...
    GoogleAuthRequest:
      type: object
      required:
        - idToken
      properties:
        idToken:
          type: string
          example: eyJhbGciOiJSUzI1NiIsImtpZCI6IjFlOWdkazcifQ...
    RefreshTokenRequest:
      type: object
      required:
        - refreshToken
      properties:
        refreshToken:
          type: string
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
...
paths:
...
  /auth/google:
    post:
      operationId: googleAuth
      summary: Authenticate with Google
      description: Authenticates a user with Google OAuth
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GoogleAuthRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /auth/refresh:
    post:
      operationId: refreshToken
      summary: Refresh access token
      description: Refreshes the access token using a refresh token
      tags:
        - Auth
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RefreshTokenRequest'
      responses:
        '200':
          description: Token refreshed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid refresh token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /auth/me:
    get:
      operationId: getCurrentUser
      summary: Get current user
      description: Returns the current authenticated user's profile
      tags:
        - Auth
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Current user retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Current user retrieved successfully
                  data:
                    $ref: '#/components/schemas/User'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
security:
  - bearerAuth: [] 